After gblues' change, the rest of the button function callbacks for the other joypad drivers still had int16_t as the return type instead of int32_t - fix this

+++ input/drivers_joypad/ctr_joypad.c
@@ -58,7 +58,7 @@ static void *ctr_joypad_init(void *data)
    return (void*)-1;
 }
 
-static int16_t ctr_joypad_button(unsigned port_num, uint16_t joykey)
+static int32_t ctr_joypad_button(unsigned port_num, uint16_t joykey)
 {
    if (port_num >= DEFAULT_MAX_PADS)
       return 0;
