+++ input/drivers_joypad/switch_joypad.c
@@ -94,7 +94,7 @@ static void *switch_joypad_init(void *data)
    return (void*)-1;
 }
 
-static int16_t switch_joypad_button(unsigned port_num, uint16_t joykey)
+static int32_t switch_joypad_button(unsigned port_num, uint16_t joykey)
 {
    if (port_num >= DEFAULT_MAX_PADS)
       return 0;
